[style🎨] : File > CommonFile 엔티티명 변경
