+++ src/main/java/shop/gaship/gashipshoppingmall/commonfile/controller/CommonFileController.java
@@ -1,12 +1,10 @@
-package shop.gaship.gashipshoppingmall.file.controller;
+package shop.gaship.gashipshoppingmall.commonfile.controller;
 
 import java.io.IOException;
-import java.net.MalformedURLException;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.io.Resource;
-import org.springframework.core.io.UrlResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -14,8 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.gaship.gashipshoppingmall.file.entity.File;
-import shop.gaship.gashipshoppingmall.file.service.FileService;
+import shop.gaship.gashipshoppingmall.commonfile.service.CommonFileService;
 
 /**
  * 파일 컨트롤러입니다.
@@ -27,13 +24,13 @@ import shop.gaship.gashipshoppingmall.file.service.FileService;
 @RequestMapping("/api/files")
 @RequiredArgsConstructor
 @Slf4j
-public class FileController {
-    private final FileService fileService;
+public class CommonFileController {
+    private final CommonFileService commonFileService;
 
     @GetMapping("/download/{fileNo}")
     public ResponseEntity<Resource> fileDownload(@PathVariable Integer fileNo,
                                                  HttpServletRequest request) {
-        Resource resource = fileService.loadResource(fileNo);
+        Resource resource = commonFileService.loadResource(fileNo);
 
         String contentType = "application/octet-stream";
         try {
